% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stadiums.R
\docType{data}
\name{stadiums}
\alias{stadiums}
\title{Geolocation data for SWAC-affiliated football stadiums}
\format{
A tibble with 28 rows and 3 variables:
\describe{
\item{stadium}{chr Name of stadium, with city and state}
\item{lat}{num Latitude of stadium, expressed in degrees north of the equator}
\item{lon}{num Longitude of stadium, with negative numbers showing degrees west of the prime meridian}
}
}
\usage{
stadiums
}
\description{
Table showing coordinates in latitude and longitude for the football stadiums used most commonly by SWAC members since 2015. Data includes coordinates for all stadiums visited at least 4 times.
}
\keyword{datasets}
