% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swac-package.R
\docType{package}
\name{swac-package}
\alias{swac}
\alias{swac-package}
\title{swac: Southwest Athletic Conference (SWAC) Data}
\description{
Data on current and former SWAC institutions, including membership, sports records, location, metadata, and lyrics of alma maters.
}
\author{
\strong{Maintainer}: James Clawson \email{clawson@gmail.com}

}
\keyword{internal}
