% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swac_almamaters.R
\docType{data}
\name{swac_almamaters}
\alias{swac_almamaters}
\title{Lyrics of alma maters from current and former SWAC member institutions}
\format{
A tibble with 18 rows and 2 variables:
\describe{
\item{university}{chr Name of institution}
\item{text}{chr Lyrics of alma mater, with line and stanza separations included}
}
}
\source{
\url{https://en.wikipedia.org/wiki/Southwestern_Athletic_Conference}
}
\usage{
swac_almamaters
}
\description{
Table of text lyrics from 18 songs. Data is suitable for simple text analysis projects.
}
\keyword{datasets}
