% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/universities.R
\docType{data}
\name{universities}
\alias{universities}
\title{Data about current and former members of the Southwestern Athletic Conference}
\format{
A tibble with 18 rows and 13 variables:
\describe{
\item{id}{chr Institutional ID assigned by U.S. Department of Education}
\item{name}{chr Institutional name}
\item{abbrev}{chr Abbreviated name for convenience}
\item{founded}{num Year of founding}
\item{city}{chr City location of institution}
\item{state}{chr State location of institution}
\item{formerly}{List Nested list showing former locations and names of institution. Each row contains a tibble with variable rows and 4 columns: former_city, former_name, from, until. These last two columns indicate years the institution was located or known by former values before moving, renaming, or both.}
\item{mascot}{chr Designation of team mascot. Where differentiated, second name indicates preferred feminine form.}
\item{colors}{List Nested list showing colors affiliated with the institution. Each row contains a tibble with variable rows (between 2 and 4) and a single column showing a color definition in hexadecimal code.}
\item{url}{chr URL to institutional website}
\item{note}{chr Any further explanations, where needed}
}
}
\usage{
universities
}
\description{
Table showing information on each of the 18 institutions that are or have been members of SWAC.
}
\keyword{datasets}
